Issue #25: Disable timezone handling on the invoice date.
UTC timezone handling was enabled in a previous fix, but this was not the right
solution. It transpired that a bug regarding timezone handling was introduced
in the latest stable release of the Date module, and using UTC time was just a
workaround that fixed the problem for 22 hours per day. We should not use
timezone handling for invoice dates obviously.